setted up the addign college functionalities correct

+++ controllers/university_controller.js
@@ -1,7 +1,8 @@
 import University from "../models/university.js";
+import cloudinary from "../config/cloudinary.js";
 
 
-export const addUniversity = async (req, res) => {
+export const addUniversity = async (req, res,next) => {
   try {
     const { name} = req.body;
     if (!req.file) {
@@ -21,15 +22,18 @@ export const addUniversity = async (req, res) => {
       );
       upload.end(req.file.buffer); // Pass file buffer to Cloudinary
     });
-    const university = new University({ name,logo: uploadResult.secure_url });
+    const university = new University({ name,logo: uploadResult.secure_url,  imagePublicId: uploadResult.public_id
+    });
     await university.save();
     res.status(201).json(university);
   } catch (err) {
+    console.log("Error uploading image:", err);
+    
     next(err);
   }
 };
 
-export const getUniversities = async (req, res) => {
+export const getUniversities = async (req, res,next) => {
   try {
     const universities = await University.find().populate("colleges");
     res.json(universities);
@@ -39,7 +43,7 @@ export const getUniversities = async (req, res) => {
 };
 
 
-export const getUniversityWithColleges = async (req, res, next) => {
+export const getUniversityWithColleges = async (req, res,next) => {
   try {
     const { id } = req.params;
 
@@ -58,7 +62,7 @@ export const getUniversityWithColleges = async (req, res, next) => {
   }
 };
 
-export const updateUniversity = async (req, res) => {
+export const updateUniversity = async (req, res,next) => {
   try {
     const { name, logo } = req.body;
 
@@ -73,11 +77,24 @@ export const updateUniversity = async (req, res) => {
   }
 };
 
-export const deleteUniversity = async (req, res) => {
+
+export const deleteUniversity = async (req, res, next) => {
   try {
+    const university = await University.findById(req.params.id);
+    if (!university) {
+      return res.status(404).json({ message: 'University not found' });
+    }
+
+    // If university has a Cloudinary public_id, delete the image
+    if (university.imagePublicId) {
+      await cloudinary.uploader.destroy(university.imagePublicId);
+    }
+
     await University.findByIdAndDelete(req.params.id);
-    res.json({ message: "University deleted" });
+
+    res.json({ message: 'University and image deleted' });
   } catch (err) {
     next(err);
   }
 };
+
